Resolve comments and add file generation
