+++ assembly/frame/index.ts
@@ -1,4 +1,4 @@
 export * from "./executive";
+export * from "./metadata";
 export * from "./models/storageEntry";
 export * from "./system";
-
