+++ assembly/runtime/runtime.ts
@@ -22,21 +22,6 @@ export type InherentType = Inherent<Moment>;
 export type ExtrinsicDataType = ExtrinsicData<ExtrinsicIndexType, ByteArray>;
 
 /**
- * @description Runtime specific methods
- */
-export namespace Runtime {
-    /**
-     * @description Metadata of the runtime
-     */
-    export function metadata(): u8[] {
-        // returns hard-coded value, currently
-        let metadata: u8[] = [
-            109,101,116,97,12,16,24,83,121,115,116,101,109,1,24,83,121,115,116,101,109,44,48,65,99,99,111,117,110,116,78,111,110,99,101,1,0,20,78,111,110,99,101,32,0,0,0,0,0,0,0,0,4,108,83,116,111,114,101,115,32,110,111,110,99,101,32,111,102,32,116,104,101,32,97,99,99,111,117,110,116,56,69,120,116,114,105,110,115,105,99,67,111,117,110,116,1,0,20,73,110,100,101,120,16,0,0,0,0,4,180,84,111,116,97,108,32,101,120,116,114,105,110,115,105,99,115,32,99,111,117,110,116,32,102,111,114,32,116,104,101,32,99,117,114,114,101,110,116,32,98,108,111,99,107,46,64,65,108,108,69,120,116,114,105,110,115,105,99,115,76,101,110,1,0,20,73,110,100,101,120,16,0,0,0,0,4,61,1,84,111,116,97,108,32,108,101,110,103,116,104,32,40,105,110,32,98,121,116,101,115,41,32,102,111,114,32,97,108,108,32,101,120,116,114,105,110,115,105,99,115,32,112,117,116,32,116,111,103,101,116,104,101,114,44,32,102,111,114,32,116,104,101,32,99,117,114,114,101,110,116,32,98,108,111,99,107,46,40,80,97,114,101,110,116,72,97,115,104,1,0,16,72,97,115,104,128,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,4,108,72,97,115,104,32,111,102,32,116,104,101,32,112,114,101,118,105,111,117,115,32,98,108,111,99,107,46,56,69,120,116,114,105,110,115,105,99,115,82,111,111,116,1,0,16,72,97,115,104,128,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,4,17,1,69,120,116,114,105,110,115,105,99,115,32,114,111,111,116,32,111,102,32,116,104,101,32,99,117,114,114,101,110,116,32,98,108,111,99,107,44,32,97,108,115,111,32,112,97,114,116,32,111,102,32,116,104,101,32,98,108,111,99,107,32,104,101,97,100,101,114,46,24,68,105,103,101,115,116,1,0,20,66,121,116,101,115,4,0,4,236,68,105,103,101,115,116,32,111,102,32,116,104,101,32,99,117,114,114,101,110,116,32,98,108,111,99,107,44,32,97,108,115,111,32,112,97,114,116,32,111,102,32,116,104,101,32,98,108,111,99,107,32,104,101,97,100,101,114,46,52,69,120,116,114,105,110,115,105,99,68,97,116,97,1,1,5,12,117,51,50,20,66,121,116,101,115,0,4,0,4,57,1,69,120,116,114,105,110,115,105,99,115,32,100,97,116,97,32,102,111,114,32,116,104,101,32,99,117,114,114,101,110,116,32,98,108,111,99,107,32,40,109,97,112,115,32,97,110,32,101,120,116,114,105,110,115,105,99,39,115,32,105,110,100,101,120,32,116,111,32,105,116,115,32,100,97,116,97,41,46,36,66,108,111,99,107,72,97,115,104,1,0,16,72,97,115,104,128,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,4,148,77,97,112,32,111,102,32,98,108,111,99,107,32,110,117,109,98,101,114,115,32,116,111,32,98,108,111,99,107,32,104,97,115,104,101,115,46,24,78,117,109,98,101,114,1,0,44,66,108,111,99,107,78,117,109,98,101,114,32,0,0,0,0,0,0,0,0,4,5,1,84,104,101,32,99,117,114,114,101,110,116,32,98,108,111,99,107,32,110,117,109,98,101,114,32,98,101,105,110,103,32,112,114,111,99,101,115,115,101,100,46,32,83,101,116,32,98,121,32,96,101,120,101,99,117,116,101,95,98,108,111,99,107,96,46,56,69,120,116,114,105,110,115,105,99,73,110,100,101,120,1,0,20,73,110,100,101,120,16,0,0,0,0,4,160,84,104,101,32,105,110,100,101,120,32,111,102,32,116,104,101,32,108,97,115,116,32,101,120,101,99,117,116,101,100,32,101,120,116,114,105,110,115,105,99,56,69,120,101,99,117,116,105,111,110,80,104,97,115,101,1,0,24,83,116,114,105,110,103,4,0,4,84,66,108,111,99,107,32,101,120,101,99,117,116,105,111,110,32,112,104,97,115,101,0,0,8,56,66,108,111,99,107,72,97,115,104,67,111,117,110,116,44,66,108,111,99,107,78,117,109,98,101,114,16,232,3,0,0,4,81,1,78,117,109,98,101,114,32,111,102,32,98,108,111,99,107,32,104,97,115,104,101,115,32,116,111,32,115,116,111,114,101,32,105,110,32,116,104,101,32,115,116,111,114,97,103,101,44,32,112,114,117,110,105,110,103,32,115,116,97,114,116,115,32,119,105,116,104,32,116,104,101,32,111,108,100,101,115,116,32,98,108,111,99,107,72,77,97,120,105,109,117,109,66,108,111,99,107,76,101,110,103,116,104,12,117,51,50,16,136,19,0,0,4,164,84,104,101,32,109,97,120,105,109,117,109,32,108,101,110,103,116,104,32,111,102,32,97,32,98,108,111,99,107,32,40,105,110,32,98,121,116,101,115,41,46,0,0,16,65,117,114,97,0,1,8,60,103,101,116,83,108,111,116,68,117,114,97,116,105,111,110,0,4,244,67,97,108,108,115,32,116,104,101,32,84,105,109,101,83,116,97,109,112,32,109,111,100,117,108,101,32,97,110,100,32,114,101,116,117,114,110,115,32,99,111,110,102,105,103,117,114,101,100,32,109,105,110,32,112,101,114,105,111,100,46,56,103,101,116,65,117,116,104,111,114,105,116,105,101,115,0,4,185,1,82,101,97,100,115,32,102,114,111,109,32,116,104,101,32,83,116,111,114,97,103,101,32,116,104,101,32,97,117,116,104,111,114,105,116,105,101,115,32,116,104,97,116,32,10,119,101,114,101,32,115,101,116,32,111,110,32,103,101,110,101,115,105,115,44,32,99,114,101,97,116,101,115,32,97,32,118,101,99,116,111,114,32,111,102,32,65,99,99,111,117,110,116,73,100,115,32,97,110,100,32,114,101,116,117,114,110,32,105,116,0,0,0,1,32,66,97,108,97,110,99,101,115,1,32,98,97,108,97,110,99,101,115,4,28,65,99,99,111,117,110,116,1,0,28,65,99,99,111,117,110,116,4,0,4,136,83,116,111,114,101,115,32,105,110,102,111,114,109,97,116,105,111,110,32,97,98,111,117,116,32,97,99,99,111,117,110,116,73,100,1,8,40,115,101,116,66,97,108,97,110,99,101,12,36,97,99,99,111,117,110,116,73,100,36,65,99,99,111,117,110,116,73,100,44,102,114,101,101,66,97,108,97,110,99,101,28,66,97,108,97,110,99,101,60,114,101,115,101,114,118,101,100,66,97,108,97,110,99,101,28,66,97,108,97,110,99,101,4,129,1,83,101,116,115,32,116,104,101,32,98,97,108,97,110,99,101,115,32,111,102,32,97,32,103,105,118,101,110,32,65,99,99,111,117,110,116,73,100,10,65,108,116,101,114,115,32,116,104,101,32,70,114,101,101,32,98,97,108,97,110,99,101,32,97,110,100,32,82,101,115,101,114,118,101,100,32,98,97,108,97,110,99,101,115,32,105,110,32,83,116,111,114,97,103,101,46,32,116,114,97,110,115,102,101,114,12,24,115,101,110,100,101,114,36,65,99,99,111,117,110,116,73,100,32,114,101,99,101,105,118,101,114,36,65,99,99,111,117,110,116,73,100,24,97,109,111,117,110,116,28,66,97,108,97,110,99,101,4,196,84,114,97,110,115,102,101,114,32,116,104,101,32,103,105,118,101,110,32,97,109,111,117,110,116,32,102,114,111,109,32,115,101,110,100,101,114,32,116,111,32,114,101,99,101,105,118,101,114,0,0,0,2,36,84,105,109,101,115,116,97,109,112,1,36,116,105,109,101,115,116,97,109,112,8,12,78,111,119,1,0,24,77,111,109,101,110,116,32,0,0,0,0,0,0,0,0,4,108,67,117,114,114,101,110,116,32,115,101,116,32,84,105,109,101,115,116,97,109,112,32,118,97,108,117,101,36,68,105,100,85,112,100,97,116,101,1,0,16,98,111,111,108,4,0,4,192,67,104,101,99,107,115,32,105,102,32,84,105,109,101,115,116,97,109,112,32,119,97,115,32,117,112,100,97,116,101,100,32,102,111,114,32,99,117,114,114,101,110,116,32,115,108,111,116,0,0,4,52,109,105,110,105,109,117,109,80,101,114,105,111,100,24,77,111,109,101,110,116,32,136,19,0,0,0,0,0,0,4,132,77,105,110,105,109,117,109,32,112,101,114,105,111,100,32,98,101,116,119,101,101,110,32,116,105,109,101,115,116,97,109,112,115,0,3,4,28,64,67,104,101,99,107,83,112,101,99,86,101,114,115,105,111,110,56,67,104,101,99,107,84,120,86,101,114,115,105,111,110,48,67,104,101,99,107,71,101,110,101,115,105,115,56,67,104,101,99,107,77,111,114,116,97,108,105,116,121,40,67,104,101,99,107,78,111,110,99,101,44,67,104,101,99,107,87,101,105,103,104,116,96,67,104,97,114,103,101,84,114,97,110,115,97,99,116,105,111,110,80,97,121,109,101,110,116,109,101,116,97,12,0]
-        return metadata;
-    }
-}
-
-/**
  * @description Constants for runtime
  */
 export class RuntimeConfig {
