--- metadata/asconfig.json
@@ -1,20 +1,0 @@
-{
-  "targets": {
-    "debug": {
-      "binaryFile": "build/untouched.wasm",
-      "textFile": "build/untouched.wat",
-      "sourceMap": true,
-      "debug": true
-    },
-    "release": {
-      "binaryFile": "build/optimized.wasm",
-      "textFile": "build/optimized.wat",
-      "sourceMap": true,
-      "optimizeLevel": 3,
-      "shrinkLevel": 1,
-      "converge": false,
-      "noAssert": false
-    }
-  },
-  "options": {}
-}
