--- metadata/assembly/tsconfig.json
@@ -1,6 +1,0 @@
-{
-  "extends": "assemblyscript/std/assembly.json",
-  "include": [
-    "./**/*.ts"
-  ]
-}
