+++ metadata/package.json
@@ -4,15 +4,10 @@
   "version": "1.0.0",
   "author": "Dastan @dastanbeksamatov",
   "dependencies": {
-    "@assemblyscript/loader": "^0.17.5",
     "@polkadot/metadata": "^2.10.1",
     "@polkadot/types": "^2.10.1",
     "@polkadot/util": "^4.2.1",
-    "assemblyscript-json": "^0.3.3",
     "typescript": "^4.1.2"
-  },
-  "devDependencies": {
-    "assemblyscript": "^0.17.5"
   },
   "homepage": "https://github.com/limechain/subsembly",
   "license": "MIT",
@@ -20,9 +15,6 @@
   "repository": "https://github.com/limechain/subsembly",
   "scripts": {
     "metadata": "node .",
-    "asbuild:untouched": "asc assembly/index.ts --target debug",
-    "asbuild:optimized": "asc assembly/index.ts --target release",
-    "asbuild": "npm run asbuild:untouched && npm run asbuild:optimized",
     "test": "node tests"
   }
 }
