+++ package.json
@@ -7,7 +7,7 @@
     "test": "asc test",
     "asbuild:untouched": "asc assembly/index.ts -b build/runtime-untouched.wasm -t build/runtime-untouched.wat --runtime half --memoryBase=16777216 --importMemory --noAssert --use abort=",
     "asbuild:optimized": "asc assembly/index.ts -b build/runtime-optimized.wasm -t build/runtime-optimized.wat --runtime half --memoryBase=16777216 --importMemory --noAssert --optimize --use abort=",
-    "build": "yarn run asbuild:optimized && node build.js",
+    "build": "yarn run metadata && yarn run asbuild:optimized && node build.js",
     "metadata": "node metadata"
   },
   "repository": {
